feature_store_ml/pyproject.toml
[tool.poetry]
name = "feature-store-ml"
version = "0.1.0"
description = "Feature store sandbox with LightGBM"
authors = ["Andrew Ghostuhin <me@torinasakura.name>"]
packages = [{include = "feature_store_ml", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
clickhouse-connect = "^0.7.6"
pandas = "^2.2"
numpy = "^1.26"
lightgbm = "^4.3"
scikit-learn = "^1.5"
pyyaml = "^6.0"
joblib = "^1.4"
loguru = "^0.7"
typer = {version = "^0.12", extras = ["all"]}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
